break job into a prepare and render phase
this, if we can add some thread-safety helpers to GdkGLContext may allow
us to build something that prepares the work on the main thread w/o any
synchronization, but render on a GL thread.

There are places where that could be really useful, like macOS where we
want to block the CVDisplayLink to perform our GL operations.