cleanup texturing code in driver
this code was a source of a lot of indirection in the previous renderer.
this tries to decode that a bit by specifying when we're caching, when
we're creating new, and when we are looking up/releasing textures for
reuse.

it will mean just a slight bit more code in renderjob, but very much work
it for the clarity of troubleshooting.