allow for multiple command queues with shared state
this hoists the attachment/uniforms up a layer from the command queue so
that we can have multiple while sharing state. Doing so allows us to use
the driver more for the top-level object and ensure that programs may
indirectly use the drivers command queue (which can change as we process
frames for each surface). When a render completes, the queue is reverted
to the shared command queue for out-of-band operations.