handle degenerate assertions in nine slices
if you are scaled way out, such as a paintable in the node-editor, you
might have a degenerate case where the assertions would be 2 < 1. In
general those cases don't matter so much because we are so far zoomed out.

This way we can leave these on in debug builds rather than having them
crash things.